stared on the user usecase
