+++ internal/pkg/user/usecase/model.go
@@ -1,0 +1,17 @@
+package usecase
+
+import (
+	"failless/internal/pkg/db"
+	"failless/internal/pkg/user"
+	"failless/internal/pkg/user/repository"
+)
+
+type UseCase struct {
+	rep user.Repository
+}
+
+func GetUseCase() UseCase {
+	return UseCase{
+		rep: repository.NewSqlUserRepository(db.ConnectToDB()),
+	}
+}
